gateway/requirements.txt
flask==2.3.3
requests==2.31.0
pyjwt==2.8.0
python-dotenv==1.0.0
flask-cors==4.0.0